+++ docs/source/dowhy.gcm.rst
@@ -118,10 +118,10 @@ dowhy.gcm.divergence module
    :undoc-members:
    :show-inheritance:
 
-dowhy.gcm.fcms module
+dowhy.gcm.causal_mechanisms module
 ---------------------
 
-.. automodule:: dowhy.gcm.fcms
+.. automodule:: dowhy.gcm.causal_mechanisms
    :members:
    :undoc-members:
    :show-inheritance:
@@ -142,10 +142,10 @@ dowhy.gcm.fitting\_sampling module
    :undoc-members:
    :show-inheritance:
 
-dowhy.gcm.graph module
+dowhy.graph module
 ----------------------
 
-.. automodule:: dowhy.gcm.graph
+.. automodule:: dowhy.graph
    :members:
    :undoc-members:
    :show-inheritance:
