+++ dowhy/gcm/validation.py
@@ -12,8 +12,8 @@ import pandas as pd
 from statsmodels.stats.multitest import multipletests
 
 from dowhy.gcm.cms import InvertibleStructuralCausalModel
-from dowhy.gcm.graph import DirectedGraph, get_ordered_predecessors, is_root_node, validate_causal_graph
 from dowhy.gcm.independence_test import kernel_based
+from dowhy.graph import DirectedGraph, get_ordered_predecessors, is_root_node, validate_causal_graph
 
 
 class RejectionResult(Enum):
